Moved security config from API to the ripple-security module
