+++ ripple-security/src/main/java/org/rippleosi/config/Pac4jConfig.java
@@ -13,7 +13,7 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-package org.rippleosi.config.security;
+package org.rippleosi.config;
 
 import org.pac4j.core.authorization.RequireAllRolesAuthorizer;
 import org.pac4j.core.authorization.RequireAnyRoleAuthorizer;
