+++ ripple-security/src/main/java/org/rippleosi/config/SecurityConfig.java
@@ -13,7 +13,7 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-package org.rippleosi.config.security;
+package org.rippleosi.config;
 
 import org.pac4j.core.config.Config;
 import org.rippleosi.security.interceptor.authentication.RequiresAuthenticationInterceptor;
